refactor(api/tokens): move v0 api back to previous location to make the with dependencies build.....
