+++ apis/tokens/api/v0/index.ts
@@ -1,6 +1,6 @@
 import type { VercelRequest, VercelResponse } from '@vercel/node'
 
-import { getTokens } from '../../lib/api/v0.js'
+import { getTokens } from '../../lib/api.js'
 // import { TokensApiSchema } from '../../lib/schemas/index.js'
 
 const handler = async (_request: VercelRequest, response: VercelResponse) => {
