+++ apis/tokens/api/v0/search/[address].ts
@@ -1,6 +1,6 @@
 import type { VercelRequest, VercelResponse } from '@vercel/node'
 
-import { getTokensByAddress } from '../../../lib/api/v0.js'
+import { getTokensByAddress } from '../../../lib/api.js'
 import { SearchTokenApiSchema } from '../../../lib/schemas/search/address.js'
 
 const handler = async (request: VercelRequest, response: VercelResponse) => {
